split messages constants on new file
