+++ src/extension/src/azure/azure-auth/azureAuth.ts
@@ -6,7 +6,9 @@ import {
   ResourceManagementClient as RMC
 } from "azure-arm-resource";
 import { AuthorizationError, ResourceGroupError } from "../../errors";
-import { CONSTANTS, DialogMessages, DialogResponses } from "../../constants";
+import { CONSTANTS } from "../../constants/constants";
+import { DialogMessages, DialogResponses } from "../../constants/dialog";
+import { MESSAGES } from "../../constants/messages";
 
 const MICROSOFT_DOCUMENT_DB_PROVIDER = "Microsoft.DocumentDb";
 const MICROSOFT_WEB_PROVIDER = "Microsoft.Web";
@@ -75,7 +77,7 @@ export abstract class AzureAuth {
       await vscode.commands.executeCommand("azure-account.login");
       // Make sure it did not return from timeout
       if (this.api.status === CONSTANTS.AZURE_LOGIN_STATUS.LOGGING_IN) {
-        throw new AuthorizationError(CONSTANTS.ERRORS.LOGIN_TIMEOUT);
+        throw new AuthorizationError(MESSAGES.ERRORS.LOGIN_TIMEOUT);
       }
       return true;
     } else {
@@ -92,7 +94,7 @@ export abstract class AzureAuth {
       await vscode.commands.executeCommand("azure-account.logout");
       // Make sure it did not return from timeout
       if (this.api.status === CONSTANTS.AZURE_LOGIN_STATUS.LOGGED_IN) {
-        throw new AuthorizationError(CONSTANTS.ERRORS.LOGOUT_FAILED);
+        throw new AuthorizationError(MESSAGES.ERRORS.LOGOUT_FAILED);
       }
     }
     return true;
@@ -178,7 +180,7 @@ export abstract class AzureAuth {
           return resourceGroup;
         }
       }
-      throw new ResourceGroupError(CONSTANTS.ERRORS.RESOURCE_GROUP_NOT_FOUND);
+      throw new ResourceGroupError(MESSAGES.ERRORS.RESOURCE_GROUP_NOT_FOUND);
     });
   }
   /*
@@ -206,7 +208,7 @@ export abstract class AzureAuth {
     subscriptionItem: SubscriptionItem
   ): Promise<LocationItem[]> {
     if (subscriptionItem === null || subscriptionItem === undefined) {
-      return Promise.reject(CONSTANTS.ERRORS.SUBSCRIPTION_NOT_DEFINED);
+      return Promise.reject(MESSAGES.ERRORS.SUBSCRIPTION_NOT_DEFINED);
     }
 
     await this.initializeLocations(subscriptionItem);
@@ -240,7 +242,7 @@ export abstract class AzureAuth {
     subscriptionItem: SubscriptionItem
   ): Promise<LocationItem[]> {
     if (subscriptionItem === null || subscriptionItem === undefined) {
-      return Promise.reject(CONSTANTS.ERRORS.SUBSCRIPTION_NOT_DEFINED);
+      return Promise.reject(MESSAGES.ERRORS.SUBSCRIPTION_NOT_DEFINED);
     }
 
     await this.initializeLocations(subscriptionItem);
